refactor: convert ops.py to package and extract self-contained sections
Convert src/boring_semantic_layer/ops.py to a package (ops/) and extract
two cleanly-bounded chunks into submodules:

- ops/_values.py (166 lines): Dimension, Measure value objects plus
  the prefix proxies that support model-prefix navigation in joined
  dimension lambdas. Imports nothing from the relational core; the only
  callback to _core (_format_column_error in the AttributeError path)
  is a lazy in-method import to avoid circular module loading.

- ops/_column_extraction.py (369 lines): ColumnTracker,
  ColumnExtractionResult, JoinColumnExtractionResult,
  TableColumnRequirements and the projection-pushdown extraction
  helpers. Self-contained except for one lazy _unwrap import inside
  _extract_requirements_from_measures.

ops/__init__.py re-exports every name external callers reference, so
all existing `from boring_semantic_layer.ops import X` imports keep
working unchanged. _core.py imports the moved names back so its own
internal call sites are unchanged too.

The remaining 4,860 lines in _core.py are the relational op classes
(Semantic*Op) and their tightly-coupled helpers (root-model traversal,
join-key/rename machinery, aggregation planning). Those resist a clean
split: helpers reference SemanticTableOp which is defined mid-file,
and the join logic threads through several intricate fixes documented
in MEMORY.md (multi-way join column ambiguity, _RenamedResolver,
pre-aggregation join direction). Splitting them further would risk
subtle behavior changes for limited structural gain. A future pass
could pull these apart but it warrants its own focused branch.

No behavior change. 930 tests pass.
